TEMPORARY HACKS, to be properly handled later:
As it turns out, certain bits of the VES2.H and ENGINE.C code aren't
the same in Duke3D 1.4's Build Editor EXE as they are in the game EXE.
It looks like the editor's revision is a bit later, albeit
significantly closer to the 1.4 game EXE than to 1.5.

Work should still be done on stereo handlers.