OK, this one has been confusing:
It hasn't been clear if Duke3D 1.4's DUKE3D.EXE uses a newer ENGINE.OBJ
revision than BUILD.EXE or not. In certain ways, this appears to be the
case, say with the grid size changes in draw2grid. BUILD.EXE's revision
is also missing the calls to _disable/_enable from the stereo handlers.
For both examples, the Duke3D 1.4 EXE is closer to Duke3D 1.5.

On the other hand, this does not hold for VES2.H:qlimitrate.
Duke3D 1.4's Build Editor and Duke3D 1.5 both use the same
implementation, while Duke3D 1.4's implementation differs.

As for the changes to the voxel rendering code in ENGINE.C, basically,
it's not always clear which revision has the more up-to-date code.
Voxel rendering was still under development back then.

For now the choice has been that the Duke3D 1.4 game EXE is using
a newer revision than the editor. The game's revision will still
be identified as 19960820, while the editor's will be 19960724,
using the Build Engine notes from BUILD2.TXT as a base again.

An exception should be made for VES2.h:qlimitrate. Maybe
a separate copy/branch of the engine was originally used
for the Duke3D game with an older VES2.H file.