ENGINE.C: Hacks for making drawalls and initengine match the March 96 revision in size. Again, various permutations of definitions were checked. There's also a bit of inline ASM for initengine.
The generated code for drawalls itself is now looking less
like the original, but at least it's matching in size.

The situation is not as bad in initengine, although this somewhat
increases the amount of differences again. For reference, there's
a bit of code left out which at least makes the usages of
local variables before the switch match the original OBJ file.

Currently only drawmasks differs in size (and also in other manners).
After the following functions of drawmaskwall and drawsprite,
a jump table is aligned on a 2-bytes boundary in such a way that
all following code more-or-less matches the original OBJ file again.