So, Duke3D v1.3d's Build Editor also uses a revision of ENGINE.OBJ somewhat differing from 19960320, possibly an earlier one. Thus, it's been decided to at least try to cover the 19960213 revision.
It is the case that after disabling all code under the SUPERBUILD macro
(voxel rendering), the 19960320 revision isn't very different from
19960213. The 95 revision of ENGINE.C, and other files, could also
be used as a reference for filling in missing gaps again.

Known remaining differences in recreated 19960213 revision code:
- wallscan: Swapped local vars and a few instances of a different
register being used; The compiled code still has the same size.
- transmaskvline: Basically the same, although the function's epilogue
lacks a jump to drawmaskwall's epilogue. Might be related to differing
size of drawmasks.
- initengine: Minor differences, as with the 19960320 revision.
A bit less, though, presumably due to the lack of
voxel-related code. Still matching in size.
- drawmasks: Minor difference with subtraction of 256, also present in
recreation of the 19960320 revision. Differs in size from the original
by 1 byte.
- drawmaskwall: Basically the same, albeit with an added jump from
its epilogue to transmaskvline's epilogue. Again possibly related
to drawmasks' size.
- completemirror: Not even bothering to list, since the recreation
of the 19960320 revision is already known to be imperfect.