Trying to recover ENGINE.OBJ for Duke3D v1.3d's Build Editor now. Generally, when there's a difference between the 19960213 and 19960320 revisions, it looks like there's a match with the former, at least in size. This might be related to the differing copybufbyte pragmas.
setvesa and setgamemode are both are same as in 19960320.

SUPERBUILD is defined; There's somewhat earlier voxel-related
code which should still be investigated, especially in drawvox
and (probably also) drawsprite.

There's a minor weirdness with pushmove, like there were an additional
unused function argument; Work should be done on completemirror, at
least partially, since the code in the EXE appears to match 19960213.